P0: harden production code, add CI, fix clippy, cache render texture
- Add GitHub Actions CI workflow (test, clippy, fmt, doc, audit)
- Add rustfmt.toml and clippy.toml configuration
- Fix panics in renderer/text.rs: propagate Typst/font errors via Result
- Cache render texture in window.rs (eliminates per-frame GPU alloc)
- Fix all clippy warnings workspace-wide (-D warnings clean)
- Add TextCompilation variant to RenderError

+++ crates/animatix/src/timeline/svg.rs
@@ -111,10 +111,10 @@ pub fn parse_svg(svg_data: &str) -> Result<Vec<VelloPath>, String> {
                         if let usvg::Paint::Color(c) = stroke.paint() {
                             stroke_opts = Some((
                                 Color::from_rgba8(c.red, c.green, c.blue, stroke.opacity().to_u8()),
-                                stroke.width().get() as f32,
+                                stroke.width().get(),
                             ));
                         } else {
-                            stroke_opts = Some((Color::BLACK, stroke.width().get() as f32));
+                            stroke_opts = Some((Color::BLACK, stroke.width().get()));
                         }
                     }
